requirements.txt
boto3>=1.17.0
python-dotenv>=0.19.0
pylint>=2.8.0
mypy>=0.812
build>=0.7.0
twine>=3.4.1
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.5b2